pyproject.toml
[tool.poetry]
name = "resume-builder"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">0.115.6"
pydantic = "^2.10.4"
django-db = "^0.0.7"
python-decouple = "^3.8"
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
